Separate UI and uploader with pubsub-js
