+++ frontend/package.json
@@ -22,6 +22,7 @@
     "webpack": "^2.2.1"
   },
   "dependencies": {
-    "crypto-js": "^3.1.9-1"
+    "crypto-js": "^3.1.9-1",
+    "pubsub-js": "^1.5.5"
   }
 }
